refactor: Remove unused EvanTestCase, move UniqueFaker to faker.py
EvanTestCase was dead code - it had been replaced by the pytest-style
OTISClient in fixtures.py. Moved UniqueFaker (which is still actively
used in factory files) to a new faker.py file and deleted testcase.py.
