+++ otisweb_testsuite/__init__.py
@@ -5,7 +5,7 @@ This package provides test helpers and base classes for Django testing,
 with enhanced assertion methods and debugging capabilities.
 """
 
-from .testcase import EvanTestCase, UniqueFaker
+from .faker import UniqueFaker
 
-__all__ = ["EvanTestCase", "UniqueFaker"]
+__all__ = ["UniqueFaker"]
 __version__ = "1.0.0"
